refactors file structure and Period class
