--- app/jobs/slack_notify_job.rb
@@ -1,9 +1,0 @@
-class SlackNotifyJob < ApplicationJob
-  queue_as :default
-
-  def perform(*args)
-    #binding.pry
-    SlackNotifier.notify_hook(args[:hook], args[:image_url], args[:message])# if params[:hook] && Rails.env.development?
-
-  end
-end
