--- spec/jobs/slack_notify_job_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe SlackNotifyJob, type: :job do
-  pending "add some examples to (or delete) #{__FILE__}"
-end
